Javascript Basics Exercises
Histogram:

Given n integers in the interval [1…1000]. Of these, some percentage p1 are below 200, another percentage p2 are from 200 to 399, another percentage p3 are from 400 to 599, another percentage p4 are from 600 to 799 and the remaining p5 percentage are from 800 and above. To write a program that calculates and prints the percentages p1, p2, p3, p4 and p5.