Javascript Basics Exercises
Oscars:
You are invited by the academy to write software to calculate the points for an actor/actress. The academy will give you initial points for the actor.Each rater will then give their rating. The points the actor receives are formed by: the length of the evaluator's name multiplied by the points he gives divided by two.
If the score at any point exceeds 1250.5 the program should abort and print that the given actor has received a nomination.